Add test for random_seed attribute and make fixes to ensure it behaves the same as the command line options
