+++ Source/VCGeneration/Checker.cs
@@ -46,6 +46,7 @@ namespace Microsoft.Boogie
     private volatile CheckerStatus status;
     private readonly CheckerPool pool;
     public volatile Program Program;
+    private readonly ProverOptions solverOptions;
 
     public void GetReady()
     {
@@ -104,18 +105,18 @@ namespace Microsoft.Boogie
       this.pool = pool;
       this.Program = prog;
 
-      ProverOptions options = cce.NonNull(CommandLineOptions.Clo.TheProverFactory).BlankProverOptions();
+      solverOptions = cce.NonNull(CommandLineOptions.Clo.TheProverFactory).BlankProverOptions();
 
       if (logFilePath != null)
       {
-        options.LogFilename = logFilePath;
+        solverOptions.LogFilename = logFilePath;
         if (appendLogFile)
         {
-          options.AppendLogFile = appendLogFile;
+          solverOptions.AppendLogFile = appendLogFile;
         }
       }
 
-      options.Parse(CommandLineOptions.Clo.ProverOptions);
+      solverOptions.Parse(CommandLineOptions.Clo.ProverOptions);
 
       ContextCacheKey key = new ContextCacheKey(prog);
       ProverInterface prover;
@@ -132,13 +133,13 @@ namespace Microsoft.Boogie
       {
         ctx = (ProverContext) cce.NonNull(ctx).Clone();
         prover = (ProverInterface)
-          CommandLineOptions.Clo.TheProverFactory.SpawnProver(CommandLineOptions.Clo, options, ctx);
+          CommandLineOptions.Clo.TheProverFactory.SpawnProver(CommandLineOptions.Clo, solverOptions, ctx);
       }
       else
       {
         if (ctx == null)
         {
-          ctx = (ProverContext) CommandLineOptions.Clo.TheProverFactory.NewProverContext(options);
+          ctx = (ProverContext) CommandLineOptions.Clo.TheProverFactory.NewProverContext(solverOptions);
         }
 
         Setup(prog, ctx, split);
@@ -147,7 +148,7 @@ namespace Microsoft.Boogie
         // context in the cache, so that the prover can setup stuff in
         // the context to be cached
         prover = (ProverInterface)
-          CommandLineOptions.Clo.TheProverFactory.SpawnProver(CommandLineOptions.Clo, options, ctx);
+          CommandLineOptions.Clo.TheProverFactory.SpawnProver(CommandLineOptions.Clo, solverOptions, ctx);
         cachedContexts.Add(key, cce.NonNull((ProverContext) ctx.Clone()));
       }
 
@@ -184,26 +185,20 @@ namespace Microsoft.Boogie
     {
       TheoremProver.SetRlimit(Util.BoundedMultiply(rlimit, 1000));
     }
-
-    private void SetRandomSeed(int? randomSeed)
-    {
-      TheoremProver.SetRandomSeed(randomSeed);
-    }
-
     
     /// <summary>
     /// Set up the context.
     /// </summary>
-    private void Setup(Program prog, ProverContext ctx, Split s = null)
+    private void Setup(Program prog, ProverContext ctx, Split split = null)
     {
-      var implementationRandomSeed = s?.Implementation.RandomSeed;
-      var random = implementationRandomSeed == null ? CommandLineOptions.Clo.Random : new Random(implementationRandomSeed.Value);
+      solverOptions.RandomSeed = split?.Implementation.RandomSeed ?? CommandLineOptions.Clo.RandomSeed;
+      var random = solverOptions.RandomSeed == null ? null : new Random(solverOptions.RandomSeed.Value);
       
       Program = prog;
       // TODO(wuestholz): Is this lock necessary?
       lock (Program.TopLevelDeclarations)
       {
-        var declarations = s == null ? prog.TopLevelDeclarations : s.TopLevelDeclarations;
+        var declarations = split == null ? prog.TopLevelDeclarations : split.TopLevelDeclarations;
         var reorderedDeclarations = GetReorderedDeclarations(declarations, random);
         foreach (var declaration in reorderedDeclarations) {
           Contract.Assert(declaration != null);
@@ -339,7 +334,7 @@ namespace Microsoft.Boogie
       }
     }
 
-    public void BeginCheck(string descriptiveName, VCExpr vc, ProverInterface.ErrorHandler handler, uint timeout, uint rlimit, int? randomSeed)
+    public void BeginCheck(string descriptiveName, VCExpr vc, ProverInterface.ErrorHandler handler, uint timeout, uint rlimit)
     {
       Contract.Requires(descriptiveName != null);
       Contract.Requires(vc != null);
@@ -358,7 +353,6 @@ namespace Microsoft.Boogie
       }
       SetTimeout(timeout);
       SetRlimit(rlimit);
-      SetRandomSeed(randomSeed);
       proverStart = DateTime.UtcNow;
       thmProver.BeginCheck(descriptiveName, vc, handler);
       //  gen.ClearSharedFormulas();    PR: don't know yet what to do with this guy
@@ -386,297 +380,6 @@ namespace Microsoft.Boogie
   // -----------------------------------------------------------------------------------------------
   // -----------------------------------------------------------------------------------------------
   // -----------------------------------------------------------------------------------------------
-
-  public abstract class ProverInterface
-  {
-    public static ProverInterface CreateProver(Program prog, string /*?*/ logFilePath, bool appendLogFile, uint timeout,
-      int taskID = -1)
-    {
-      Contract.Requires(prog != null);
-
-      ProverOptions options = cce.NonNull(CommandLineOptions.Clo.TheProverFactory).BlankProverOptions();
-
-      if (logFilePath != null)
-      {
-        options.LogFilename = logFilePath;
-        if (appendLogFile)
-        {
-          options.AppendLogFile = appendLogFile;
-        }
-      }
-
-      if (timeout > 0)
-      {
-        options.TimeLimit = Util.BoundedMultiply(timeout, 1000);
-      }
-
-      if (taskID >= 0)
-      {
-        options.Parse(CommandLineOptions.Clo.Cho[taskID].ProverOptions);
-      }
-      else
-      {
-        options.Parse(CommandLineOptions.Clo.ProverOptions);
-      }
-
-      ProverContext ctx = (ProverContext) CommandLineOptions.Clo.TheProverFactory.NewProverContext(options);
-
-      // set up the context
-      foreach (Declaration decl in prog.TopLevelDeclarations)
-      {
-        Contract.Assert(decl != null);
-        TypeCtorDecl t = decl as TypeCtorDecl;
-        if (t != null)
-        {
-          ctx.DeclareType(t, null);
-        }
-      }
-
-      foreach (Declaration decl in prog.TopLevelDeclarations)
-      {
-        Contract.Assert(decl != null);
-        Constant c = decl as Constant;
-        if (c != null)
-        {
-          ctx.DeclareConstant(c, c.Unique, null);
-        }
-        else
-        {
-          Function f = decl as Function;
-          if (f != null)
-          {
-            ctx.DeclareFunction(f, null);
-          }
-        }
-      }
-
-      foreach (var ax in prog.Axioms)
-      {
-        ctx.AddAxiom(ax, null);
-      }
-
-      foreach (Declaration decl in prog.TopLevelDeclarations)
-      {
-        Contract.Assert(decl != null);
-        GlobalVariable v = decl as GlobalVariable;
-        if (v != null)
-        {
-          ctx.DeclareGlobalVariable(v, null);
-        }
-      }
-
-      return (ProverInterface) CommandLineOptions.Clo.TheProverFactory.SpawnProver(CommandLineOptions.Clo, options, ctx);
-    }
-
-    public enum Outcome
-    {
-      Valid,
-      Invalid,
-      TimeOut,
-      OutOfMemory,
-      OutOfResource,
-      Undetermined,
-      Bounded
-    }
-
-    public readonly ISet<VCExprVar> NamedAssumes = new HashSet<VCExprVar>();
-    public ISet<string> UsedNamedAssumes { get; protected set; }
-
-    public class ErrorHandler
-    {
-      // Used in CheckOutcomeCore
-      public virtual int StartingProcId()
-      {
-        return 0;
-      }
-
-      public virtual void OnModel(IList<string> labels, Model model, Outcome proverOutcome)
-      {
-        Contract.Requires(cce.NonNullElements(labels));
-      }
-
-      public virtual void OnResourceExceeded(string message,
-        IEnumerable<Tuple<AssertCmd, TransferCmd>> assertCmds = null)
-      {
-        Contract.Requires(message != null);
-      }
-
-      public virtual void OnProverWarning(string message)
-      {
-        Contract.Requires(message != null);
-        switch (CommandLineOptions.Clo.PrintProverWarnings)
-        {
-          case CommandLineOptions.ProverWarnings.None:
-            break;
-          case CommandLineOptions.ProverWarnings.Stdout:
-            Console.WriteLine("Prover warning: " + message);
-            break;
-          case CommandLineOptions.ProverWarnings.Stderr:
-            Console.Error.WriteLine("Prover warning: " + message);
-            break;
-          default:
-            Contract.Assume(false);
-            throw new cce.UnreachableException(); // unexpected case
-        }
-      }
-
-      public virtual void OnProverError(string message)
-      {
-        // no-op by default.
-        //Errors are always printed to console by the prover
-      }
-
-      public virtual Absy Label2Absy(string label)
-      {
-        Contract.Requires(label != null);
-        Contract.Ensures(Contract.Result<Absy>() != null);
-
-        throw new System.NotImplementedException();
-      }
-    }
-
-    public abstract void BeginCheck(string descriptiveName, VCExpr vc, ErrorHandler handler);
-
-    [NoDefaultContract]
-    public abstract Outcome CheckOutcome(ErrorHandler handler, int errorLimit);
-
-    public virtual string[] CalculatePath(int controlFlowConstant)
-    {
-      throw new System.NotImplementedException();
-    }
-
-    public virtual void LogComment(string comment)
-    {
-      Contract.Requires(comment != null);
-    }
-
-    public virtual void Close()
-    {
-    }
-
-    public abstract void Reset(VCExpressionGenerator gen);
-
-    public abstract void FullReset(VCExpressionGenerator gen);
-
-    /// <summary>
-    /// MSchaef: Allows to Push a VCExpression as Axiom on the prover stack (beta)
-    /// for now it is only implemented by ProcessTheoremProver and still requires some
-    /// testing
-    /// </summary>
-    public virtual void PushVCExpression(VCExpr vc)
-    {
-      Contract.Requires(vc != null);
-      throw new NotImplementedException();
-    }
-
-    public virtual string VCExpressionToString(VCExpr vc)
-    {
-      Contract.Requires(vc != null);
-      Contract.Ensures(Contract.Result<string>() != null);
-      throw new NotImplementedException();
-    }
-
-    public virtual void Pop()
-    {
-      Contract.EnsuresOnThrow<UnexpectedProverOutputException>(true);
-      throw new NotImplementedException();
-    }
-
-    public virtual int NumAxiomsPushed()
-    {
-      throw new NotImplementedException();
-    }
-
-    public virtual int FlushAxiomsToTheoremProver()
-    {
-      Contract.EnsuresOnThrow<UnexpectedProverOutputException>(true);
-      throw new NotImplementedException();
-    }
-
-    // (assert vc)
-    public virtual void Assert(VCExpr vc, bool polarity, bool isSoft = false, int weight = 1, string name = null)
-    {
-      throw new NotImplementedException();
-    }
-
-    public virtual List<string> UnsatCore()
-    {
-      throw new NotImplementedException();
-    }
-
-
-    // (assert implicit-axioms)
-    public virtual void AssertAxioms()
-    {
-      throw new NotImplementedException();
-    }
-
-    // (check-sat)
-    public virtual void Check()
-    {
-      throw new NotImplementedException();
-    }
-
-    // (check-sat + get-unsat-core + checkOutcome)
-    public virtual Outcome CheckAssumptions(List<VCExpr> assumptions, out List<int> unsatCore, ErrorHandler handler)
-    {
-      throw new NotImplementedException();
-    }
-
-    public virtual Outcome CheckAssumptions(List<VCExpr> hardAssumptions, List<VCExpr> softAssumptions,
-      out List<int> unsatisfiedSoftAssumptions, ErrorHandler handler)
-    {
-      throw new NotImplementedException();
-    }
-
-    public virtual Outcome CheckOutcomeCore(ErrorHandler handler, int taskID = -1)
-    {
-      throw new NotImplementedException();
-    }
-
-    // (push 1)
-    public virtual void Push()
-    {
-      throw new NotImplementedException();
-    }
-
-    // Set theorem prover timeout for the next "check-sat"
-    public virtual void SetTimeout(uint ms)
-    {
-    }
-
-    public virtual void SetRlimit(uint limit)
-    {
-    }
-
-    public virtual void SetRandomSeed(int? randomSeed)
-    {
-    }
-
-    public abstract ProverContext Context { get; }
-
-    public abstract VCExpressionGenerator VCExprGen { get; }
-
-    public virtual void DefineMacro(Macro fun, VCExpr vc)
-    {
-      throw new NotImplementedException();
-    }
-
-    public class VCExprEvaluationException : Exception
-    {
-    }
-
-    public virtual object Evaluate(VCExpr expr)
-    {
-      throw new NotImplementedException();
-    }
-
-    // Assert vc tagged with a name
-    public virtual void AssertNamed(VCExpr vc, bool polarity, string name)
-    {
-      throw new NotImplementedException();
-    }
-  }
 
   public class ProverInterfaceContracts : ProverInterface
   {
